LayoutTests/fast/dom/Range/getBoundingClientRect-expected.txt

Test 1
PASS rect.left.toFixed(3) is "8.000"
PASS rect.top.toFixed(3) is "8.000"
PASS rect.width.toFixed(3) is "400.000"
PASS rect.height.toFixed(3) is "400.000"
PASS rect.right is rect.left + rect.width
PASS rect.bottom is rect.top + rect.height

Test 2
PASS rect.left.toFixed(3) is "8.000"
FAIL rect.top.toFixed(3) should be 452.000. Was 453.000.
PASS rect.width.toFixed(3) is "400.000"
PASS rect.height.toFixed(3) is "376.000"
PASS rect.right is rect.left + rect.width
PASS rect.bottom is rect.top + rect.height

Test 3
PASS rect.left.toFixed(3) is "8.000"
FAIL rect.top.toFixed(3) should be 1044.000. Was 1045.000.
PASS rect.width.toFixed(3) is "400.000"
PASS rect.height.toFixed(3) is "96.000"
PASS rect.right is rect.left + rect.width
PASS rect.bottom is rect.top + rect.height

Test 4
PASS rect.left.toFixed(3) is "0.000"
PASS rect.top.toFixed(3) is "0.000"
PASS rect.width.toFixed(3) is "0.000"
PASS rect.height.toFixed(3) is "0.000"
PASS rect.right is rect.left + rect.width
PASS rect.bottom is rect.top + rect.height

Test 5
FAIL rect.left.toFixed(3) should be -14.574. Was -15.340.
FAIL rect.top.toFixed(3) should be 1761.947. Was 1762.590.
PASS rect.width.toFixed(3) is "504.009"
PASS rect.height.toFixed(3) is "535.849"
PASS Math.abs(rect.left + rect.width - rect.right) < 0.001 is true
PASS Math.abs(rect.top + rect.height - rect.bottom) < 0.001 is true

Test 6
PASS rect.left.toFixed(3) is "0.000"
PASS rect.top.toFixed(3) is "0.000"
PASS rect.width.toFixed(3) is "0.000"
PASS rect.height.toFixed(3) is "0.000"
PASS rect.right is rect.left + rect.width
PASS rect.bottom is rect.top + rect.height

Test 7
PASS rect.left.toFixed(3) is "8.000"
FAIL rect.top.toFixed(3) should be 2180.000. Was 2181.000.
PASS rect.width.toFixed(3) is "176.000"
PASS rect.height.toFixed(3) is "96.000"
PASS rect.right is rect.left + rect.width
PASS rect.bottom is rect.top + rect.height

PASS successfullyParsed is true

TEST COMPLETE
